requirements.txt
aiogram==3.17.0
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
folium==0.19.5
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
pyfiglet==1.0.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.4.0
xyzservices==2025.4.0